Assignment2

+++ src/edu/dmacc/coma510/Problem1.java
@@ -2,12 +2,12 @@ package edu.dmacc.coma510;
 
 public class Problem1 {
     public static void main(String[] args) {
-        
-        int numberOfPlateAppearances = 100;
-        int numberOfWalks = 4;
+
+        double numberOfPlateAppearances = 100;
+        double numberOfWalks = 4;
         int numberOfHits = 24;
 
-        int calculatedBattingAverage = numberOfHits / (numberOfPlateAppearances - numberOfWalks);
+        double calculatedBattingAverage = numberOfHits / (numberOfPlateAppearances - numberOfWalks);
 
         System.out.println(calculatedBattingAverage); // 0.25
     }
